Attach the instances to the load balancer.
The name of resources or modules provisioned with count refers to the entire collection. In this example, aws_instance.app now refers to all of the EC2 instances. You can reference individual items in collections with the same notation as list indexing. For example, aws_instance.app[0] refers to the first instance Terraform provisions.
You can create a list of all of the values of a given attribute for the items in the collection with an asterik. For instance, aws_instance.app.*.id will be a list of all the IDs of the instances.